Use PIT for rollup jobs as well
